requirements.txt
streamlit==0.74.1
pandas==1.0.5
numpy==1.18.5
scikit-learn==0.23.1
pillow==7.2.0
